Agrego usuarioServivio y UsuarioControler, queda operativa la API usuario/Get.

+++ Modelo/Usuario.cs
@@ -8,13 +8,13 @@ namespace Modelo
 {
     public class Usuario
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nombre { get; set; }
         public string Apellido { get; set; }
         public string Correo { get; set; }
         public string Usr { get; set; }
         public string Pwd { get; set; }
-        public int IdPerfil { get; set; }
+        public long IdPerfil { get; set; }
         public string Avatar { get; set; }
         public bool Sexo { get; set; }
         public DateTime FechaNacimiento { get; set; }
